requirements.txt
blinker==1.7.0
click==8.1.7
confluent-kafka==2.3.0
Faker==24.11.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
kafka-python==2.0.2
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.2
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
tzdata==2024.1
watchdog==4.0.0
Werkzeug==3.0.1
